+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on <https://common-changelog.org/>, and this project adheres mostly to Semantic Versioning. However, all releases before 1.0.0 have breaking changes between minor-version updates.
 
+## [2.1.0] - 2024-07-06
+
+### Added
+
+- Implemented trigger rules for social events
+- Added `OnSocialEvent` event to `SocialEngine` and `SocialEngineController` that's invoked whenever a social event is dispatched.
+
 ## [2.0.0] - 2024-02-28
 
 Version 2.0.0's architecture has been heavily refactored on the back-end to simplify testing, reliability, and serialization (save/load). Users will find a similar experience within the inspector, but some core content authoring features have been moved around.
@@ -63,3 +70,4 @@ _Initial release._
 [1.0.0]: https://github.com/ShiJbey/TDRS/releases/tag/v1.0.0
 [1.1.0]: https://github.com/ShiJbey/TDRS/releases/tag/v1.1.0
 [2.0.0]: https://github.com/ShiJbey/TDRS/releases/tag/v2.0.0
+[2.1.0]: https://github.com/ShiJbey/TDRS/releases/tag/v2.1.0